Start peer with state DOWN initially
Start with peer as DOWN initially and then make them NORMAL after
conenction is established.
A NORMAL state means a node is up and connected.
Self node is NORMAL by default.
Peer node becomes normal after a connection is successful.
Also make sure the peer is marked DOWN as soon as the connection breaks
so that the higher layers can react accordingly.

This will help in rerouting traffic to a different rack as soon as the
node in the remote region goes down