Remote DC replication Failover
This is the final patch in the series to enable remote DC replication
failover. If a node figures a remote node is marked as DOWN, it tries to
failover to another node. This does not failover those requests which
were already queued on the connection and could not be forwarded due to
network issues, but only those that arrived after the remote peer was
marked as DOWN.