magic number in object to detect corruptions
Every object has a magic number in the beginning which is used during
printing to detect any pointer corruptions.
There is not need for the switch case of types. The value within the
range along with the magic number is a good enough check.
Move the printf registration from logs to types since its on an object
now.