Fix memory corruption with peer array
We create a peers array in the server pool which was a array of struct
node. The address of this struct was being handed all over which is can
become invalid when the array resizes and reallocs the underlying
memory. The solution is to have arrays as a array of pointers and hand
over those pointers.

There is a need for a special type of array which will hold only
pointers.